==> celltype_annotation/analysis/02-cell-assign-sarcoma-marker-genes.Rmd <==
---
title: "CellAssign: Rhabdomyosarcoma using marker genes"
output: 
  html_notebook:
    toc: true
    toc_float: true
params:
  library_id: "SCPCL000488"
  sample_id: "SCPCS000262"
---

Since it's fairly obvious that `CellAssign`, although better at assigning `NA` than `SingleR`, is still going to be highly dependent on the marker genes that are in the reference, I wanted to see how well `CellAssign` performs when we provide marker genes that we know are associated with the cell types found in the dataset.

Here I use `scran::FindMarkers()` to get the top markers for each of the cell types and then use that as the reference gene by cell matrix for `CellAssign`.
I expect that the output of `CellAssign` should be fairly consistent with the annotations we have using this method.

## Set Up

```{r}
library(SingleCellExperiment)
library(ggplot2)
```

```{r}
# Set up paths 

# build path to annotated sce file 
processed_sce_file <- glue::glue("{params$library_id}_processed_celltype.rds")
processed_sce_path <- file.path("..", "data", params$sample_id, processed_sce_file)

# output matrix files 
rms_matrix_file <- file.path("..", "references", "rms_broad_markers_mtx.csv")

# anndata input
anndata_dir <- file.path("..", "data", "anndata")
anndata_file <- glue::glue("{params$library_id}_anndata.hdf5")
anndata_file <- file.path(anndata_dir, anndata_file)

# prediction output
cellassign_outs_dir <- "cellassign_results"
fs::dir_create(cellassign_outs_dir)

# cellassign predictions
rms_prediction_file <- file.path(
  cellassign_outs_dir, 
  glue::glue("{params$library_id}_rms_broad_markers.tsv")
)

```

```{r}
# read in all annotated sce 
processed_sce <- readr::read_rds(processed_sce_path)
```

```{r}
# source helper functions
function_file <- file.path("..", "utils", "cellassign-helper-functions.R")
source(function_file)
```

## Get marker genes

Below we will find the upregulated marker genes when comparing each cell type to all other cell types. 
The cell types here were obtained from the submitter.
The original submitter cell types contained sub groups of tumor cells, e.g., `Tumor_myoblast-A` and `Tumor_myoblast-B`. 
Before finding markers, the subgroups are combined into a single cell type, e.g., `Tumor_myoblast`.
These cell types are then used to group cells when applying `scran::FindMarkers()`.
Then the markers are sorted by logFC and the top 10 are selected.
This ensures that we only get marker genes that are upregulated in a specified cell type compared to all other cells in the dataset.
Because of this, we are only identifying marker genes that are unique to each cell type and are not considering marker genes that may be expressed in a subset of cell types.

```{r}
# extract coldata for easy plotting
coldata_df <- colData(processed_sce) |>
  as.data.frame() |> 
  dplyr::select(barcode, celltype) |>
  # classify cells as tumor or normal
  dplyr::mutate(cell_category = dplyr::if_else(stringr::str_detect(celltype, "Tumor"), "Tumor", celltype),
                # remove sub states of myoblast/mesoderm/myocyte
                broad_celltype = stringr::str_remove(celltype, "-A|-B|-C|-D"))
```


```{r}
# find markers for all cell types 
processed_sce_markers <- scran::findMarkers(processed_sce,
                                            groups = coldata_df$broad_celltype,
                                            pval.type = "all")
```

Below the top 10 upregulated markers for each cell type are converted into a binary gene by cell matrix to use as input for `CellAssign`.

```{r}
# get top 10 markers for each cell type 
top_markers_df <- as.list(processed_sce_markers) |>
  purrr::imap(
    \(markers, id) 
    markers |> 
      as.data.frame() |>
      dplyr::arrange(desc(summary.logFC)) |>
      dplyr::mutate(celltype = id) |>
      tibble::rownames_to_column("ensembl_id") |> 
      dplyr::select(celltype, ensembl_id) |>
      head(n = 10)
  ) |>
  dplyr::bind_rows()

# create a binary matrix of genes by cell type markers 
markers_binary_mtx <- top_markers_df |>
  dplyr::select(celltype, ensembl_id) |> 
  tidyr::pivot_wider(ensembl_id,
                     names_from = celltype,
                     values_from = celltype,
                     values_fn = length,
                     values_fill = 0) |> 
  tibble::column_to_rownames("ensembl_id") |>
  dplyr::mutate(other = 0)

# replace length with 1
markers_binary_mtx[markers_binary_mtx > 1] <- 1

# add ensembl id as first column
markers_binary_mtx <- markers_binary_mtx |>
  tibble::rownames_to_column("ensembl_id") |>
  dplyr::relocate(ensembl_id)
```

```{r}
# look at myoblast markers, are they what we expect? 
myoblast_markers <- top_markers_df |>
  dplyr::filter(celltype == "Tumor_Myoblast")

# do some mapping to get the gene symbols
myoblast_markers <- AnnotationDbi::mapIds(
    # organism annotation package
    org.Hs.eg.db::org.Hs.eg.db,
    # the provided gene identifiers
    keys = myoblast_markers$ensembl_id,
    # the type of provided gene identifiers
    keytype = "ENSEMBL",
    # the type of gene identifiers to map to
    column = "SYMBOL"
  ) |>
    tibble::enframe(name = 'ensembl_id',
                    value = 'symbol') |> 
    dplyr::left_join(myoblast_markers, by = "ensembl_id")

myoblast_markers
```

I kind of would expect some of the myoblast markers to be included here like `MYOG`, `MYF5`, `MYOD1`, etc. 
Perhaps these genes are expressed in other cell types, such as normal myoblast cells that may be present. 
If they are expressed at a high enough level in other cell types then this strategy of "all" used here for `FindMarkers` would not identify those genes. 
Running the same test using "any" however, did not reveal these genes either, or at least they were not in the top 10 (data not shown). 
Another approach could be to filter by a logFC cutoff rather than just taking the top 10.

## Run CellAssign
```{r}
# export mtx 
readr::write_csv(markers_binary_mtx, rms_matrix_file)
```

```{r,  eval = FALSE}
# run cell assign with muscle ref
cellassign_call <- glue::glue("
python ../scripts/cell-assign.py \
  --input_anndata '{anndata_file}' \
  --output_predictions '{rms_prediction_file}' \
  --reference '{rms_matrix_file}'
")

system(cellassign_call)
```

## Evaluate CellAssign

```{r}
# read in prediction scores 
rms_predictions <- readr::read_tsv(rms_prediction_file)
```

```{r}
# heatmap of prediction scores 
rms_predictions |>
  dplyr::select(-barcode) |> 
  as.matrix() |> 
  pheatmap::pheatmap(show_rownames = FALSE)
```

This actually looks pretty good! 
We see most of the cells are assigned to `Tumor_Myoblast` which is what is expected.

```{r}
# add rms marker gene assignments to sce object
rms_assignments <- get_celltype_assignments(rms_predictions)
coldata_df$rms_assignments <- rms_assignments$celltype
```

```{r}
# compare to broad celltype annotations
compare_refs_heatmap(original_assignment = coldata_df$broad_celltype,
                     cell_assign = coldata_df$rms_assignments)
```

Although there is not perfect agreement, we do see that for many of the cell types the majority of cells are classified as the expected cell type.

## Session Info

```{r}
sessionInfo()
```
